Measure time in milliseconds with precision of 3 points after comma

+++ main.c
@@ -23,7 +23,7 @@ const char* const INPUT_FILENAMES[12] = {
     "inputs/day_12.txt",
 };
 uint64_t solutions[24] = {0};
-uint64_t durations_in_ms[24] = {0};
+double durations_in_ms[24] = {0};
 
 void solve(int8_t day, int8_t part) {
     fprintf(stdout, "Solving Day %d Part %d... ", day, part);
@@ -39,7 +39,7 @@ void solve(int8_t day, int8_t part) {
         solutions[index] = SOLVE_PFNS[index](&puzzle_input);
         struct timespec end;
         clock_gettime(CLOCK_MONOTONIC, &end);
-        durations_in_ms[index] = (end.tv_sec - start.tv_sec) * 1000 + (end.tv_nsec - start.tv_nsec) / 1000000;
+        durations_in_ms[index] = (end.tv_sec - start.tv_sec) * 1000 + (double)(end.tv_nsec - start.tv_nsec) / 1000000;
 
         destroy_puzzle_input(&puzzle_input);
         fprintf(stdout, "\r");
@@ -81,7 +81,7 @@ void print_all_solutions() {
         }
         printf("|\n");
         for (int8_t i = 0; i < 4; ++i) {
-            printf("|        | %16ld ms ", durations_in_ms[2 * (4 * j + i)]);
+            printf("|        | %16.3f ms ", durations_in_ms[2 * (4 * j + i)]);
         }
         printf("|\n");
         for (int8_t i = 0; i < 4; ++i) {
@@ -89,7 +89,7 @@ void print_all_solutions() {
         }
         printf("|\n");
         for (int8_t i = 0; i < 4; ++i) {
-            printf("|        | %16ld ms ", durations_in_ms[2 * (4 * j + i) + 1]);
+            printf("|        | %16.3f ms ", durations_in_ms[2 * (4 * j + i) + 1]);
         }
         printf("|\n");
         for (int8_t i = 0; i < 4; ++i) {
